// src/interfaces/Iusuario.ts
export interface Iusuario {
  idUsuario?: number;
  nome: string;
  cpf: string;
  ra: string;
  dataNascimento: Date;
  genero: boolean; // true = masculino, false = feminino
  cep: string;
  endereco: string;
  numero: string;
  cidade: string;
  estado: string;
  email: string;
  telefone: string;
  senha: string;
  tipoUsuario: "passageiro" | "motorista";
  cnh?: string; // somente para motoristas
}

export interface IusuarioFiltros {
  nome?: string;
  email?: string;
  tipoUsuario?: "passageiro" | "motorista";
}

export interface IRetornoCadastroUsuario {
  id: number;
}


  